fix: type identifier fields as integers, not floats
Numeric ids (id, *_id) were typed 'number' in the spec, so the generated
Rust models used f64 and JSON serialized them as floats (1873345.0). Add
an integer_id_fields() normalizer pass that retypes id-like number fields
to integer/int64; genuine decimals (balance, *_cost) keep their type
because their names do not match. Regenerated src/models accordingly.

+++ src/models/add_cluster_to_project_request.rs
@@ -16,11 +16,11 @@ use crate::models;
 pub struct AddClusterToProjectRequest {
     /// ID добавляемого кластера.
     #[serde(rename = "resource_id")]
-    pub resource_id: f64
+    pub resource_id: i64
 }
 
 impl AddClusterToProjectRequest {
-    pub fn new(resource_id: f64) -> AddClusterToProjectRequest {
+    pub fn new(resource_id: i64) -> AddClusterToProjectRequest {
         AddClusterToProjectRequest {
             resource_id
         }
