+++ src/models/app_disk_status.rs
@@ -24,7 +24,7 @@ pub struct AppDiskStatus {
     pub size:    Option<f64>,
     /// ID диска.
     #[serde(rename = "disk_id", skip_serializing_if = "Option::is_none")]
-    pub disk_id: Option<f64>
+    pub disk_id: Option<i64>
 }
 
 impl AppDiskStatus {
