+++ src/models/create_storage_request_configurator.rs
@@ -21,7 +21,7 @@ pub struct CreateStorageRequestConfigurator {
     pub disk: Option<f64>,
     /// ID конфигуратора хранилища.
     #[serde(rename = "id", skip_serializing_if = "Option::is_none")]
-    pub id:   Option<f64>
+    pub id:   Option<i64>
 }
 
 impl CreateStorageRequestConfigurator {
