+++ src/models/dns_record.rs
@@ -25,7 +25,7 @@ pub struct DnsRecord {
         with = "::serde_with::rust::double_option",
         skip_serializing_if = "Option::is_none"
     )]
-    pub id:     Option<Option<f64>>,
+    pub id:     Option<Option<i64>>,
     #[serde(rename = "data")]
     pub data:   Box<models::DnsRecordData>,
     /// Время жизни DNS-записи.
